move classes to a utils file, move to a dir, fix stuff
